Create Output Files by Time Series
Again, this will assume you have a time ordered event files (I named mine EventsTS.csv) with which to work. Below are the UNIX instructions to order the original Events.csv file:

Open a terminal window in the folder where the Events.csv file is located.
In the terminal window use the following commands-

head –n1 Events.csv > EventsTS.csv
tail –n+2 Events.csv | sort –t’,’ –k10,10 >> EventsTS.csv
 
The head command above with the –n1 option just saves out the column labels into the new file.
 
The tail command with the special option –n+2 starts “tailing” the file at line 2 (thus skipping the labels). This is then piped (the “|” character) into the sort command.
 
The >> means append to the file, so the output of the sort is added after the labels which are already in the file Events.csv